feat: Add clients panel page and table component
The code changes include adding a new page called "ClientsPanel" and a table component called "TableClients". This allows users to view a list of clients and their details in a table format. The "ClientsPanel" page also includes a heading and a total count of clients.

This commit message follows the convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.